Title: 학점 계산기, Time: 152ms, Memory: 12MB, Status: Passed - Codetree

+++ 251126/학점 계산기/credit-calculator.java
@@ -5,14 +5,14 @@ public class Main {
        Scanner sc = new Scanner(System.in);
 
        int n = sc.nextInt();
-       int sum = 0;
+       double sum = 0;
        double avg = 0;
 
        for(int i = 0; i < n; i++){
         double sub = sc.nextDouble();
         sum += sub;
        }
-        avg = (double)sum/n;
+        avg = sum/n;
         System.out.printf("%.1f \n", avg);
 
         if(avg >= 4.0){
